+++ backend/app/models/club.rb
@@ -27,6 +27,7 @@ class Club < ApplicationRecord
   has_many :aplicacion_planes,  class_name: 'AplicacionPlan', dependent: :destroy
   has_many :pesajes_manicura,   class_name: 'PesajeManicura', dependent: :destroy
   has_many :stocks,             dependent: :destroy
+  has_many :rutas_entrega,      class_name: 'RutaEntrega', dependent: :destroy
   has_many :webhooks,           dependent: :destroy
   has_many :turnos,             dependent: :destroy
   has_many :check_ins,          dependent: :destroy